backend/requirements.txt
pandas
faker
numpy
flask
sqlalchemy
xgboost
shap
fpdf2
scipy
